API -  Get NowPlaying and Upcoming Films
